requirements.txt
torch
numpy
matplotlib
pandas
PyYAML
websockets
tensorflow
